## 유동현 ##
불 - 편한 build 경고 전부 지우고 widget 틀 올림.

+++ AutoChess/AutoChess/AutoChess.cpp
@@ -27,5 +27,5 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance,
 	WPARAM wParam = window->Run();
 	SafeDelete(window);
 
-	return wParam;
+	return static_cast<int>(wParam);
 }
